+++ packages/components/molecules/f-rating/src/components/Rating.vue
@@ -2,13 +2,14 @@
     <div
         :class="$style['c-rating']"
         data-test-id="rating-component">
-        <div :class="$style['c-rating-starWrapper']">
-            <div :class="$style['c-rating-container']">
+        <div :class="$style['c-rating-stars']">
+            <div :class="$style['c-rating-stars-icons']">
                 <star-icon
                     v-for="star in maxStarRating"
                     :key="star"
                     :class="[
-                        $style['c-rating-star-empty'],
+                        $style['c-rating-star'],
+                        $style['c-rating-star--empty'],
                         $style[`c-rating-star--${starRatingSize}`]
                     ]" />
             </div>
@@ -16,15 +17,15 @@
             <div
                 :class="[
                     $style['c-rating-mask'],
-                    $style['c-rating-container']
+                    $style['c-rating-stars-icons']
                 ]"
-                :style="`--starRatingPercentage: ${getRatingStarPercentage}`">
+                :style="`--starRatingPercentage: ${getRatingStarPercentage}%`">
                 <star-filled-icon
                     v-for="star in maxStarRating"
                     :key="star"
-                    :style="`--starRatingSize: c-rating-star--${starRatingSize}`"
                     :class="[
-                        $style['c-rating-star-filled'],
+                        $style['c-rating-star'],
+                        $style['c-rating-star--filled'],
                         $style[`c-rating-star--${starRatingSize}`]
                     ]" />
             </div>
@@ -35,6 +36,15 @@
                 {{ getRatingDescription }}
             </span>
         </div>
+
+        <span
+            v-if="ratingDisplayType"
+            data-test-id="c-rating-displayType"
+            :class="[
+                $style['c-rating-message'],
+                $style[`c-rating-message--${ratingDisplayType}`]]">
+            {{ getRatingDisplayFormat() }}
+        </span>
     </div>
 </template>
 
@@ -45,7 +55,10 @@ import {
 } from '@justeattakeaway/pie-icons-vue';
 import { VueGlobalisationMixin } from '@justeat/f-globalisation';
 import tenantConfigs from '../tenants';
-import { VALID_STAR_RATING_SIZES } from '../constants';
+import {
+    VALID_STAR_RATING_SIZES,
+    VALID_STAR_RATING_DISPLAY_TYPE
+} from '../constants';
 
 export default {
     name: 'VRating',
@@ -74,6 +87,15 @@ export default {
             type: String,
             default: 'small',
             validator: value => !!VALID_STAR_RATING_SIZES[value]
+        },
+        ratingDisplayType: {
+            type: String,
+            default: null,
+            validator: value => VALID_STAR_RATING_DISPLAY_TYPE.includes(value)
+        },
+        reviewCount: {
+            type: Number,
+            default: null
         }
     },
 
@@ -94,12 +116,12 @@ export default {
             return this.starRating < 2
                 ? this.$tc('ratings.starsDescription', 1, {
                     rating: this.starRating,
-                    total: this.maxStarRating
+                    maxStarRating: this.maxStarRating
                 })
 
                 : this.$tc('ratings.starsDescription', 2, {
                     rating: this.starRating,
-                    total: this.maxStarRating
+                    maxStarRating: this.maxStarRating
                 });
         },
 
@@ -109,7 +131,25 @@ export default {
          * @returns {string}
          */
         getRatingStarPercentage () {
-            return `${(this.starRating / this.maxStarRating) * 100}%`;
+            return (this.starRating / this.maxStarRating) * 100;
+        }
+    },
+
+    methods: {
+        /**
+         * Gets the correct rating display format from translations.
+         *
+         * @todo - If the component is using `short` as a `starRatingSize` we shouldn't display text
+         * alongside it for now. (TBC with design - ticket in backlog).
+         *
+         * @returns {string}
+         */
+        getRatingDisplayFormat () {
+            return this.$t(`ratings.ratingDisplayType.${this.ratingDisplayType}`, {
+                rating: this.starRating,
+                maxStarRating: this.maxStarRating,
+                reviewCount: this.reviewCount
+            });
         }
     }
 };
@@ -118,37 +158,39 @@ export default {
 <style lang="scss" module>
 @use '@justeat/fozzie/src/scss/fozzie' as f;
 
-.c-rating-starWrapper {
+.c-rating {
+    display: flex;
+}
+
+.c-rating-stars {
     margin: 0;
     padding: 0;
     position: relative;
     display: inline-block;
 }
-    .c-rating-container {
+    .c-rating-stars-icons {
         display: flex;
     }
 
     .c-rating-star {
-        &--small {
-            width: 12px;
-        }
-
-        &--medium {
-            width: 16px;
-        }
-
-        &--large {
-            width: 28px;
-        }
+        width: 12px;
     }
 
-    .c-rating-star-filled {
+    .c-rating-star--medium {
+        width: 16px;
+    }
+
+    .c-rating-star--large {
+        width: 28px;
+    }
+
+    .c-rating-star--filled {
         & path {
             fill: f.$color-support-brand-01;
         }
     }
 
-    .c-rating-star-empty {
+    .c-rating-star--empty {
         & path {
             fill: f.$color-mozzarella-50;
         }
@@ -164,5 +206,13 @@ export default {
         svg {
             flex-shrink: 0;
         }
+    }
+
+    .c-rating-message {
+        display: flex;
+        align-items: center;
+        @include f.font-size('body-s');
+        font-weight: f.$font-weight-bold;
+        margin-left: f.spacing(a);
     }
 </style>