+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v7.39.0
+ ------------------------------
+ *October 24, 2022*
+
+ ### Removed
+ - Duplicate dependencies of `molecules` packages.
+
+
+v7.38.0
+------------------------------
+*October 24, 2022*
+
+### Removed
+- Duplicate dependencies of `atoms` packages.
+
 
 v7.37.1
 ------------------------------